Check for Numeric, not float
ruby decodes json integers into integer types, so don't assume
that they're floats

+++ lib/jwt/verify.rb
@@ -43,7 +43,7 @@ module JWT
     def verify_iat
       return unless @payload.include?('iat')
 
-      if !(@payload['iat'].is_a?(Float)) || @payload['iat'].to_f > (Time.now.to_f + leeway)
+      if !(@payload['iat'].is_a?(Numeric)) || @payload['iat'].to_f > (Time.now.to_f + leeway)
         fail(JWT::InvalidIatError, 'Invalid iat')
       end
     end
